Game stats to generate reward from and potential non-image state
