--- scripts/envrionments/Pokemon.py
@@ -1,404 +1,0 @@
-import sys
-import uuid 
-import os
-from math import floor, sqrt
-import json
-from pathlib import Path
-
-import mediapy as media
-
-from pyboy import PyBoy
-from pyboy.utils import WindowEvent
-
-import logging
-
-import cv2
-
-import numpy as np
-from collections import deque
-
-# from typing import override
-from functools import cached_property
-
-from util.configurations import GymEnvironmentConfig
-
-class Pokemon:
-    def __init__(self, config: GymEnvironmentConfig) -> None:
-        logging.info(f"Training with Task {config.task}")
-        headless = False
-        self.task = config.task
-
-        self.rom_path = f'{Path.home()}/cares_rl_configs/pokemon/PokemonRed.gb'
-        self.init_path = f'{Path.home()}/cares_rl_configs/pokemon/init.state'
-        #config.rom_path
-
-        self.valid_actions = [
-            WindowEvent.PRESS_ARROW_DOWN,
-            WindowEvent.PRESS_ARROW_LEFT,
-            WindowEvent.PRESS_ARROW_RIGHT,
-            WindowEvent.PRESS_ARROW_UP,
-            WindowEvent.PRESS_BUTTON_A,
-            WindowEvent.PRESS_BUTTON_B,
-        ]
-        
-        # self.valid_actions.extend([
-        #     WindowEvent.PRESS_BUTTON_START,
-        #     WindowEvent.PASS
-        # ])
-
-        self.release_button = [
-            WindowEvent.RELEASE_ARROW_DOWN,
-            WindowEvent.RELEASE_ARROW_LEFT,
-            WindowEvent.RELEASE_ARROW_RIGHT,
-            WindowEvent.RELEASE_ARROW_UP,
-            WindowEvent.RELEASE_BUTTON_A,
-            WindowEvent.RELEASE_BUTTON_B
-        ]
-
-        self.act_freq = 24
-
-        head, hide_window = ['headless', True] if headless else ['SDL2', False]
-        self.pyboy = PyBoy(
-                self.rom_path,
-                debugging=False,
-                disable_input=False,
-                window_type=head,
-                hide_window=hide_window,
-            )
-
-        self.agent_stats = []
-
-        with open(self.init_path, "rb") as f:
-            self.pyboy.load_state(f)
-
-        self.screen = self.pyboy.botsupport_manager().screen()
-
-    @cached_property
-    def min_action_value(self):
-        return 0
-    
-    @cached_property
-    def max_action_value(self):
-        return len(self.valid_actions)
-
-    @cached_property
-    def observation_space(self):
-        raise NotImplementedError() # TODO figure out exactly what our observation space is - note we will have an image based version of this class
-    
-    @cached_property
-    def action_num(self):
-        return len(self.valid_actions)
-
-    def set_seed(self, seed):
-        self.seed = seed # There isn't a random element to set
-
-    def reset(self):
-        # restart game, skipping credits
-        with open(self.init_state, "rb") as f:
-            self.pyboy.load_state(f)
-
-    def grab_frame(self, height=240, width=300):
-        frame = self.screen.screen_ndarray()
-        frame = cv2.resize(frame, (width, height))
-        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB) # Convert to BGR for use with OpenCV
-        return frame
-
-    def step(self, action):
-        
-        self.run_action_on_emulator(action)
-        
-        self.append_agent_stats(action)
-
-        new_state = self._generate_state()
-        reward = self._calculate_reward(new_state)
-        done = self._check_if_done()
-        self.state = new_state
-      
-        return self.state, reward, done, False # for consistency with open ai gym just add false for truncated
-    
-    def run_action_on_emulator(self, action):
-        # press button then release after some steps - enough to move 
-        self.pyboy.send_input(self.valid_actions[action])
-        for i in range(self.act_freq):
-            self.pyboy.tick()
-            if i == 8: # ticks required to carry a "step" in the world
-              self.pyboy.send_input(self.release_button[action])
-
-    def _generate_state(self):
-        pass
-      
-    def _calculate_reward(self, new_state):
-        pass
-    
-    def _check_if_done(self):
-        return False
-
-    def append_agent_stats(self, action):
-        x_pos = self.read_m(0xD362)
-        y_pos = self.read_m(0xD361)
-        map_n = self.read_m(0xD35E)
-        levels = [self.read_m(a) for a in [0xD18C, 0xD1B8, 0xD1E4, 0xD210, 0xD23C, 0xD268]]
-    
-        self.agent_stats.append({
-            'x': x_pos, 'y': y_pos, 'map': map_n,
-            'map_location': self.get_map_location(map_n),
-            'last_action': action,
-            'pcount': self.read_m(0xD163), 
-            'levels': levels, 
-            'levels_sum': sum(levels),
-            'ptypes': self.read_party(),
-            'hp': self.read_hp_fraction(),
-            'badge': self.get_badges(),
-        })
-        print(self.agent_stats)
-
-    def update_reward(self):
-        # compute reward
-        old_prog = self.group_rewards()
-        self.progress_reward = self.get_game_state_reward()
-        new_prog = self.group_rewards()
-        new_total = sum([val for _, val in self.progress_reward.items()]) #sqrt(self.explore_reward * self.progress_reward)
-        new_step = new_total - self.total_reward
-        if new_step < 0 and self.read_hp_fraction() > 0:
-            #print(f'\n\nreward went down! {self.progress_reward}\n\n')
-            self.save_screenshot('neg_reward')
-    
-        self.total_reward = new_total
-        return (new_step, 
-                   (new_prog[0]-old_prog[0], 
-                    new_prog[1]-old_prog[1], 
-                    new_prog[2]-old_prog[2])
-               )
-    
-    def group_rewards(self):
-        prog = self.progress_reward
-        # these values are only used by memory
-        return (prog['level'] * 100 / self.reward_scale, 
-                self.read_hp_fraction()*2000, 
-                prog['explore'] * 150 / (self.explore_weight * self.reward_scale))
-               #(prog['events'], 
-               # prog['levels'] + prog['party_xp'], 
-               # prog['explore'])
-
-    def read_m(self, addr):
-        return self.pyboy.get_memory_value(addr)
-
-    def read_bit(self, addr, bit: int) -> bool:
-        # add padding so zero will read '0b100000000' instead of '0b0'
-        return bin(256 + self.read_m(addr))[-bit-1] == '1'
-    
-    def get_levels_sum(self):
-        poke_levels = [max(self.read_m(a) - 2, 0) for a in [0xD18C, 0xD1B8, 0xD1E4, 0xD210, 0xD23C, 0xD268]]
-        return max(sum(poke_levels) - 4, 0) # subtract starting pokemon level
-    
-    def get_levels_reward(self):
-        explore_thresh = 22
-        scale_factor = 4
-        level_sum = self.get_levels_sum()
-        if level_sum < explore_thresh:
-            scaled = level_sum
-        else:
-            scaled = (level_sum-explore_thresh) / scale_factor + explore_thresh
-        self.max_level_rew = max(self.max_level_rew, scaled)
-        return self.max_level_rew
-    
-    def get_knn_reward(self):
-        
-        pre_rew = self.explore_weight * 0.005
-        post_rew = self.explore_weight * 0.01
-        cur_size = self.knn_index.get_current_count() if self.use_screen_explore else len(self.seen_coords)
-        base = (self.base_explore if self.levels_satisfied else cur_size) * pre_rew
-        post = (cur_size if self.levels_satisfied else 0) * post_rew
-        return base + post
-    
-    def get_badges(self):
-        return self.bit_count(self.read_m(0xD356))
-
-    def read_party(self):
-        return [self.read_m(addr) for addr in [0xD164, 0xD165, 0xD166, 0xD167, 0xD168, 0xD169]]
-    
-    def update_heal_reward(self):
-        cur_health = self.read_hp_fraction()
-        # if health increased and party size did not change
-        if (cur_health > self.last_health and
-                self.read_m(0xD163) == self.party_size):
-            if self.last_health > 0:
-                heal_amount = cur_health - self.last_health
-                if heal_amount > 0.5:
-                    print(f'healed: {heal_amount}')
-                    self.save_screenshot('healing')
-                self.total_healing_rew += heal_amount * 4
-            else:
-                self.died_count += 1
-                
-    def get_all_events_reward(self):
-        # adds up all event flags, exclude museum ticket
-        event_flags_start = 0xD747
-        event_flags_end = 0xD886
-        museum_ticket = (0xD754, 0)
-        base_event_flags = 13
-        return max(
-            sum(
-                [
-                    self.bit_count(self.read_m(i))
-                    for i in range(event_flags_start, event_flags_end)
-                ]
-            )
-            - base_event_flags
-            - int(self.read_bit(museum_ticket[0], museum_ticket[1])),
-        0,
-    )
-
-    def get_game_state_reward(self, print_stats=False):
-        # addresses from https://datacrystal.romhacking.net/wiki/Pok%C3%A9mon_Red/Blue:RAM_map
-        # https://github.com/pret/pokered/blob/91dc3c9f9c8fd529bb6e8307b58b96efa0bec67e/constants/event_constants.asm
-        '''
-        num_poke = self.read_m(0xD163)
-        poke_xps = [self.read_triple(a) for a in [0xD179, 0xD1A5, 0xD1D1, 0xD1FD, 0xD229, 0xD255]]
-        #money = self.read_money() - 975 # subtract starting money
-        seen_poke_count = sum([self.bit_count(self.read_m(i)) for i in range(0xD30A, 0xD31D)])
-        all_events_score = sum([self.bit_count(self.read_m(i)) for i in range(0xD747, 0xD886)])
-        oak_parcel = self.read_bit(0xD74E, 1) 
-        oak_pokedex = self.read_bit(0xD74B, 5)
-        opponent_level = self.read_m(0xCFF3)
-        self.max_opponent_level = max(self.max_opponent_level, opponent_level)
-        enemy_poke_count = self.read_m(0xD89C)
-        self.max_opponent_poke = max(self.max_opponent_poke, enemy_poke_count)
-        
-        if print_stats:
-            print(f'num_poke : {num_poke}')
-            print(f'poke_levels : {poke_levels}')
-            print(f'poke_xps : {poke_xps}')
-            #print(f'money: {money}')
-            print(f'seen_poke_count : {seen_poke_count}')
-            print(f'oak_parcel: {oak_parcel} oak_pokedex: {oak_pokedex} all_events_score: {all_events_score}')
-        '''
-        
-        state_scores = {
-            'event': self.reward_scale*self.update_max_event_rew(),  
-            #'party_xp': self.reward_scale*0.1*sum(poke_xps),
-            'level': self.reward_scale*self.get_levels_reward(), 
-            'heal': self.reward_scale*self.total_healing_rew,
-            'op_lvl': self.reward_scale*self.update_max_op_level(),
-            'dead': self.reward_scale*-0.1*self.died_count,
-            'badge': self.reward_scale*self.get_badges() * 5,
-            #'op_poke': self.reward_scale*self.max_opponent_poke * 800,
-            #'money': self.reward_scale* money * 3,
-            #'seen_poke': self.reward_scale * seen_poke_count * 400,
-            'explore': self.reward_scale * self.get_knn_reward()
-        }
-        
-        return state_scores
-        
-    def update_max_op_level(self):
-        #opponent_level = self.read_m(0xCFE8) - 5 # base level
-        opponent_level = max([self.read_m(a) for a in [0xD8C5, 0xD8F1, 0xD91D, 0xD949, 0xD975, 0xD9A1]]) - 5
-        #if opponent_level >= 7:
-        #    self.save_screenshot('highlevelop')
-        self.max_opponent_level = max(self.max_opponent_level, opponent_level)
-        return self.max_opponent_level * 0.2
-    
-    def update_max_event_rew(self):
-        cur_rew = self.get_all_events_reward()
-        self.max_event_rew = max(cur_rew, self.max_event_rew)
-        return self.max_event_rew
-
-    def read_hp_fraction(self):
-        hp_sum = sum([self.read_hp(add) for add in [0xD16C, 0xD198, 0xD1C4, 0xD1F0, 0xD21C, 0xD248]])
-        max_hp_sum = sum([self.read_hp(add) for add in [0xD18D, 0xD1B9, 0xD1E5, 0xD211, 0xD23D, 0xD269]])
-        max_hp_sum = max(max_hp_sum, 1)
-        return hp_sum / max_hp_sum
-
-    def read_hp(self, start):
-        return 256 * self.read_m(start) + self.read_m(start+1)
-
-    # built-in since python 3.10
-    def bit_count(self, bits):
-        return bin(bits).count('1')
-
-    def read_triple(self, start_add):
-        return 256*256*self.read_m(start_add) + 256*self.read_m(start_add+1) + self.read_m(start_add+2)
-    
-    def read_bcd(self, num):
-        return 10 * ((num >> 4) & 0x0f) + (num & 0x0f)
-    
-    def read_money(self):
-        return (100 * 100 * self.read_bcd(self.read_m(0xD347)) + 
-                100 * self.read_bcd(self.read_m(0xD348)) +
-                self.read_bcd(self.read_m(0xD349)))
-
-    def get_map_location(self, map_idx):
-        map_locations = {
-            0: "Pallet Town",
-            1: "Viridian City",
-            2: "Pewter City",
-            3: "Cerulean City",
-            12: "Route 1",
-            13: "Route 2",
-            14: "Route 3",
-            15: "Route 4",
-            33: "Route 22",
-            37: "Red house first",
-            38: "Red house second",
-            39: "Blues house",
-            40: "oaks lab",
-            41: "Pokémon Center (Viridian City)",
-            42: "Poké Mart (Viridian City)",
-            43: "School (Viridian City)",
-            44: "House 1 (Viridian City)",
-            47: "Gate (Viridian City/Pewter City) (Route 2)",
-            49: "Gate (Route 2)",
-            50: "Gate (Route 2/Viridian Forest) (Route 2)",
-            51: "viridian forest",
-            52: "Pewter Museum (floor 1)",
-            53: "Pewter Museum (floor 2)",
-            54: "Pokémon Gym (Pewter City)",
-            55: "House with disobedient Nidoran♂ (Pewter City)",
-            56: "Poké Mart (Pewter City)",
-            57: "House with two Trainers (Pewter City)",
-            58: "Pokémon Center (Pewter City)",
-            59: "Mt. Moon (Route 3 entrance)",
-            60: "Mt. Moon",
-            61: "Mt. Moon",
-            68: "Pokémon Center (Route 4)",
-            193: "Badges check gate (Route 22)"
-        }
-        if map_idx in map_locations.keys():
-            return map_locations[map_idx]
-        else:
-            return "Unknown Location"
-
-class PoekmonImage(Pokemon):
-    def __init__(self, config: GymEnvironmentConfig, k=3):
-        self.k    = k  # number of frames to be stacked
-        self.frames_stacked = deque([], maxlen=k)
-
-        self.frame_width = 84
-        self.frame_height = 84
-
-        logging.info(f"Image Observation is on")
-        super().__init__(config)
-
-    # @override
-    @property
-    def observation_space(self):
-        return (9, self.frame_width, self.frame_height)
-
-    # @override
-    def reset(self):
-        _ = self.env.reset()
-        frame = self.grab_frame(height=self.frame_height, width=self.frame_width)
-        frame = np.moveaxis(frame, -1, 0)
-        for _ in range(self.k):
-            self.frames_stacked.append(frame)
-        stacked_frames = np.concatenate(list(self.frames_stacked), axis=0)
-        return stacked_frames
-
-    # @override
-    def step(self, action):
-        _, reward, done, truncated, _ = self.env.step(action)
-        frame = self.grab_frame(height=self.frame_height, width=self.frame_width)
-        frame = np.moveaxis(frame, -1, 0)
-        self.frames_stacked.append(frame)
-        stacked_frames = np.concatenate(list(self.frames_stacked), axis=0)
-        return stacked_frames, reward, done, truncated
